pm-internship-recommendation/backend/requirements-basic.txt
Flask==3.0.3
Flask-Cors==4.0.1
flask-sqlalchemy==3.1.1
SQLAlchemy==2.0.31
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
rapidfuzz==3.9.7
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
